main screen page controller added
main screen content is now can be swiped left or right
